made some example functions
